allow the SafeSigner towait for a safe to exist
