+++ packages/safe-tools/index.ts
@@ -1,3 +1,4 @@
 export * from "./typechain-types";
 export * from "./src/SafeRelayer"
 export * from "./src/tokenCreator"
+export * from "./src/SafeSigner"
