+++ packages/safe-tools/tests/SafeRelayer.test.ts
@@ -1,11 +1,14 @@
-import { ContractNetworksConfig } from "@safe-global/safe-core-sdk";
-import { expect } from "chai";
-import { deployments, ethers } from "hardhat";
-import { SafeRelayer } from "../src/SafeRelayer";
-import { WalletDeployer__factory } from "../typechain-types";
-const { providers, Wallet } = ethers
 
-describe("SafeSigner", () => {
+import { expect } from 'chai'
+import { ContractTransactionReceipt } from 'ethers'
+import { deployments, ethers } from 'hardhat'
+import EthersAdapter from '@safe-global/safe-ethers-lib'
+import Safe, { SafeFactory, SafeAccountConfig, ContractNetworksConfig } from '@safe-global/safe-core-sdk'
+import { getBytesAndCreateToken } from '../src/tokenCreator'
+import { EnglishOwnerAdder, EnglishOwnerRemover, WalletDeployer } from '../typechain-types'
+import { loadFixture } from "@nomicfoundation/hardhat-network-helpers"
+
+describe("SafeRelayer", () => {
     const setupTest = deployments.createFixture(
         async ({ deployments, getNamedAccounts, ethers }) => {
             await deployments.fixture(); // ensure you start from a fresh deployments
@@ -15,7 +18,12 @@ describe("SafeSigner", () => {
             const deployer = await ethers.getSigner(deployerAddress)
             const signers = await ethers.getSigners()
 
-            const chainId = await deployer.provider!.getNetwork().then(n => n.chainId)
+            const ethAdapter = new EthersAdapter({
+                ethers,
+                signerOrProvider: deployer
+            })
+
+            const chainId = await ethAdapter.getChainId()
 
             const contractNetworks: ContractNetworksConfig = {
                 [chainId]: {
@@ -28,119 +36,174 @@ describe("SafeSigner", () => {
                     createCallAddress: deploys.CreateCall.address,
                 }
             }
-            const WalletDeployerFactory:WalletDeployer__factory = await ethers.getContractFactory("WalletDeployer")
 
-            const walletDeployer = (WalletDeployerFactory).attach(deploys.WalletDeployer.address).connect(deployer)
-
-            expect(Buffer.from(WalletDeployerFactory.bytecode.slice(2), "hex").length).to.be.lessThan(28_000)
-
-            const testContractFactory = await ethers.getContractFactory("TestContract")
-            const testContract = await testContractFactory.deploy()
-            await testContract.deployed()
-
-            const relayer = new SafeRelayer({
-                ethers,
-                signer: signers[1],
-                walletDeployerAddress: walletDeployer.address,
-                EnglishOwnerAdderAddress: deploys.EnglishOwnerAdder.address,
-                networkConfig: contractNetworks,
-                provider: deployer.provider!,
-                faucet: async (address: string) => {
-                    await (await deployer.sendTransaction({
-                        to: address,
-                        value: ethers.utils.parseEther("2")
-                    })).wait()
-                }
-            })
-
-            // we need to do this within the fixture so that the safe is created before the test runs
-            // and when the state is snapshotted it all works.
-            await relayer.ready
-
-            return { deployer, signers, walletDeployer, deploys, contractNetworks, relayer, testContract, chainId }
+            const walletDeployer = (await ethers.getContractFactory("WalletDeployer")).attach(deploys.WalletDeployer.address).connect(deployer) as WalletDeployer
+            return { deployer, signers, walletDeployer, deploys, contractNetworks, ethAdapter }
         }
     );
 
-    it("executes wrapped transactions", async () => {
-        const { relayer, testContract } = await setupTest()
-        
-        const wrapped = relayer.wrappedSigner()
-        const tx = await testContract.connect(wrapped).echo("hi", false)
-        const receipt = await tx.wait()
-        
-        expect(receipt.events?.length).to.equal(1)
-        expect((receipt.events![0] as any).args.sender).to.equal((await relayer.safe)!.getAddress())
-    });
+    it("allows safe creation with sdk", async () => {
+        const { deployer, ethAdapter, contractNetworks, signers } = await setupTest()
 
-    it("finds the same safe again with a new relayer", async () => {
-        const { relayer, signers, testContract, contractNetworks, walletDeployer, deploys, deployer } = await setupTest()
+        const safeFactory = await SafeFactory.create({ ethAdapter, contractNetworks })
+        const owners = [deployer.address, signers[1].address]
+        const threshold = 1
+        const safeAccountConfig: SafeAccountConfig = {
+            owners,
+            threshold,
+            // ...
+        }
 
-        const newRelayer = new SafeRelayer({
-            ethers,
-            signer: signers[1],
-            walletDeployerAddress: walletDeployer.address,
-            EnglishOwnerAdderAddress: deploys.EnglishOwnerAdder.address,
-            networkConfig: contractNetworks,
-            provider: deployer.provider!,
-            faucet: async (address: string) => {
-                await (await deployer.sendTransaction({
-                    to: address,
-                    value: ethers.utils.parseEther("2")
-                })).wait()
-            }
-        })
-
-        const existingSafe = await relayer.safe!
-        
-        
-        expect((await newRelayer.safe)?.getAddress()).to.equal(existingSafe.getAddress())
-
-        const wrapped = relayer.wrappedSigner()
-        await expect(testContract.connect(wrapped).echo("hi", false)).to.not.be.reverted
-    });
-
-    it('executes a transaction after a fire-and-forget setSigner', async () => {
-        const { relayer, testContract } = await setupTest()
-        
-        const wrapped = relayer.wrappedSigner()
-        const tx = await testContract.connect(wrapped).echo("hi", false)
-        const receipt = await tx.wait()
-        
-        expect(receipt.events?.length).to.equal(1)
-        expect((receipt.events![0] as any).args.sender).to.equal((await relayer.safe)!.getAddress())
+        const safe: Safe = await safeFactory.deploySafe({ safeAccountConfig })
+        expect(safe).to.be.ok
+        expect(await safe.getOwners()).to.have.lengthOf(2)
+        expect(await safe.isOwner(deployer.address)).to.be.true
     })
 
-    it("reverts correctly", async () => {
-        const { relayer, testContract } = await setupTest()
-        
-        const wrapped = relayer.wrappedSigner()
-        const tx = testContract.connect(wrapped).echo("hi", true, { gasLimit: 1000000, gasPrice: 100 })
-        await expect(tx).to.be.reverted
-    });
-
-    it('executes reads on the target chain instead of the original signer chain', async () => {
-        const { testContract, walletDeployer, deploys, contractNetworks, deployer } = await setupTest()
-        // first create a default polygon chain signer
-        const provider = new providers.StaticJsonRpcProvider("https://polygon-rpc.com")
-        const signer = Wallet.createRandom().connect(provider)
-
-        const relayer = new SafeRelayer({
-            ethers,
-            signer: signer,
-            walletDeployerAddress: walletDeployer.address,
-            EnglishOwnerAdderAddress: deploys.EnglishOwnerAdder.address,
-            networkConfig: contractNetworks,
-            provider: deployer.provider!,
-            faucet: async (address: string) => {
-                await (await deployer.sendTransaction({
-                    to: address,
-                    value: ethers.utils.parseEther("2")
-                })).wait()
-            }
+    async function proxyAddressFromReceipt(receipt: ContractTransactionReceipt, ethAdapter: EthersAdapter, contractNetworks: ContractNetworksConfig) {
+        const proxyContract = ethAdapter.getSafeProxyFactoryContract({
+            safeVersion: "1.3.0",
+            chainId: await ethAdapter.getChainId(),
+            customContractAddress: contractNetworks[await ethAdapter.getChainId()].safeProxyFactoryAddress
         })
 
-        const wrapped = relayer.wrappedSigner()
-        expect(await testContract.connect(wrapped).somethingToRead()).to.equal("helloWorld")
+        const proxyCreationEvent = receipt?.events?.map((e) => {
+            try {
+                return proxyContract.contract.interface.parseLog(e)
+            } catch (err) {
+                return null
+            }
+        }).find(
+            (evt) => {
+                return evt?.name === "ProxyCreation"
+            }
+        )
+        if (!proxyCreationEvent || !proxyCreationEvent.args) {
+            console.error(proxyCreationEvent, proxyCreationEvent?.args)
+            throw new Error('SafeProxy was not deployed correctly')
+        }
+        return proxyCreationEvent.args[0]
+    }
 
+    it("uses the wallet deployer to deploy a safe", async () => {
+        const { signers, walletDeployer, ethAdapter, contractNetworks } = await setupTest()
+        const alice = signers[1]
+        const walletCreator = walletDeployer.connect(alice)
+        const aliceDevice = signers[2]
+
+        const { tokenRequest, signature } = await getBytesAndCreateToken(walletDeployer, alice, aliceDevice.address)
+        const tx = walletCreator.createSafe(tokenRequest, signature)
+        await expect(tx).to.not.be.reverted
+
+        // const safeFactory = await SafeFactory.create({ ethAdapter, contractNetworks })
+
+        const receipt = await (await tx).wait()
+        const proxyAddress = await proxyAddressFromReceipt(receipt, ethAdapter, contractNetworks)
+
+        const safe = await Safe.create({ ethAdapter, contractNetworks, safeAddress: proxyAddress })
+        expect(await safe.getOwners()).to.have.lengthOf(2)
+        expect(await safe.isOwner(aliceDevice.address)).to.be.true
     })
+
+    it("saves wallet address", async () => {
+        const { signers, walletDeployer, ethAdapter, contractNetworks } = await setupTest()
+        const alice = signers[1]
+        const walletCreator = walletDeployer.connect(alice)
+        const aliceDevice = signers[2]
+
+        const { tokenRequest, signature } = await getBytesAndCreateToken(walletDeployer, alice, aliceDevice.address)
+        const tx = walletCreator.createSafe(tokenRequest, signature)
+        await expect(tx).to.not.be.reverted
+
+        // const safeFactory = await SafeFactory.create({ ethAdapter, contractNetworks })
+
+        const receipt = await (await tx).wait()
+        const proxyAddress = await proxyAddressFromReceipt(receipt, ethAdapter, contractNetworks)
+
+        expect(await walletDeployer.ownerToSafe(alice.address)).to.equal(proxyAddress)
+        expect(await walletDeployer.safeToOwner(proxyAddress)).to.equal(alice.address)
+    })
+
+    describe("EnglishOwnerAdder", () => {
+        const fixture = async () => {
+            const { signers, deployer, deploys, walletDeployer, ethAdapter, contractNetworks } = await setupTest()
+            const alice = signers[1]
+            const aliceDevice = signers[2]
+
+            const englishOwnerAdder = (await ethers.getContractFactory("EnglishOwnerAdder")).attach(deploys.EnglishOwnerAdder.address).connect(deployer) as EnglishOwnerAdder
+            const { tokenRequest, signature } = await getBytesAndCreateToken(walletDeployer, alice, aliceDevice.address)
+
+            const receipt = await (await (walletDeployer as WalletDeployer).createSafe(tokenRequest, signature)).wait()
+
+            const proxyAddress = await proxyAddressFromReceipt(receipt, ethAdapter, contractNetworks)
+
+            const safe = await Safe.create({ ethAdapter, contractNetworks, safeAddress: proxyAddress })
+            expect(await safe.getOwners()).to.have.lengthOf(2)
+
+            return { englishOwnerAdder, signers, deployer, safe, alice, aliceDevice, walletDeployer }
+        }
+
+        it("adds the english owner adder at setup", async () => {
+            const { englishOwnerAdder, safe } = await loadFixture(fixture)
+            expect(await safe.isModuleEnabled(englishOwnerAdder.address)).to.be.true
+        })
+
+        it("can add an owner using an english signed token", async () => {
+            const { englishOwnerAdder, signers, safe, alice } = await loadFixture(fixture)
+            const newOwner = signers[3]
+            const { tokenRequest, signature } = await getBytesAndCreateToken(englishOwnerAdder, alice, newOwner.address)
+            const tx = englishOwnerAdder.addOwner(
+                safe.getAddress(),
+                tokenRequest,
+                signature
+            )
+            await expect(tx).to.not.be.reverted
+            expect(await safe.isOwner(newOwner.address)).to.be.true
+            expect(await safe.getOwners()).to.have.lengthOf(3)
+        })
+    })
+
+    describe("EnglishOwnerRemover", () => {
+        const fixture = async () => {
+            const { signers, deployer, deploys, walletDeployer, ethAdapter, contractNetworks } = await setupTest()
+            const alice = signers[1]
+            const aliceDevice = signers[2]
+
+            const englishOwnerRemover = (await ethers.getContractFactory("EnglishOwnerRemover")).attach(deploys.EnglishOwnerRemover.address).connect(deployer) as EnglishOwnerRemover
+            
+            const { tokenRequest, signature } = await getBytesAndCreateToken(walletDeployer, alice, aliceDevice.address)
+            const receipt = await (await (walletDeployer as WalletDeployer).createSafe(tokenRequest, signature)).wait()
+            const proxyAddress = await proxyAddressFromReceipt(receipt, ethAdapter, contractNetworks)
+
+            const safe = await Safe.create({ ethAdapter, contractNetworks, safeAddress: proxyAddress })
+            expect(await safe.getOwners()).to.have.lengthOf(2)
+
+            return { englishOwnerRemover, signers, deployer, safe, alice, aliceDevice, walletDeployer }
+        }
+
+        it("is added at setup", async () => {
+            const { englishOwnerRemover, safe } = await loadFixture(fixture)
+            expect(await safe.isModuleEnabled(englishOwnerRemover.address)).to.be.true
+        })
+
+        it("can remove an owner using an english signed token", async () => {
+            const { englishOwnerRemover, signers, safe, walletDeployer, alice, aliceDevice } = await loadFixture(fixture)
+            // this is a little strange because we have to find the owner right before aliceDevice in the list of owners and use that as the value for previousOwner
+            const owners = await safe.getOwners()
+            const aliceDeviceIndex = owners.findIndex((o) => o === aliceDevice.address)
+            const previousOwner = owners[Math.min(aliceDeviceIndex - 1, 0)]
+
+            const { tokenRequest, signature } = await getBytesAndCreateToken(englishOwnerRemover, alice, aliceDevice.address)
+            const tx = englishOwnerRemover.removeOwner(
+                safe.getAddress(),
+                previousOwner,
+                tokenRequest,
+                signature
+            )
+            await expect(tx).to.not.be.reverted
+            expect(await safe.isOwner(aliceDevice.address)).to.be.false
+            expect(await safe.getOwners()).to.have.lengthOf(1)
+        })
+    })
+
 })
